feat: 마이 인포에 팔로우수 추가

+++ src/main/java/com/server/objet/domain/auth/MyInfoResponseDto.java
@@ -15,14 +15,14 @@ public class MyInfoResponseDto {
     private final String profile;
 
     @Getter
-    private final Long followingArtistNum;
+    private final int followingArtistNum;
 
     @Getter
     private final String email;
 
 
     @Builder
-    public MyInfoResponseDto(String name, String profile, Long followingArtistNum, String email){
+    public MyInfoResponseDto(String name, String profile, int followingArtistNum, String email){
         this.name = name;
         this.followingArtistNum = followingArtistNum;
         this.profile = profile;
